requirements.txt

asgiref==3.3.1
channels==2.4.0
daphne==3.0.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
Jinja2==2.11.2
Pillow==5.1.0
pyinstaller-hooks-contrib==2020.7
python-apt==1.6.5+ubuntu0.2
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
simplejson==3.13.2
SQLAlchemy==1.3.16
sqlparse==0.4.
ubuntu-drivers-common==0.0.0
ufw==0.36
usb-creator==0.3.3
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
gunicorn==20.0.4